// src/renderer/types.ts
export interface Equipment {
  hunt?: {
    [name: string]: {
      name: string;
      identifier: string;
      breakFactor: number;
      minBreakFactor: number;
      hand: string;
      shield: string;
      set: string;
      slot: string;
      talent: string;
      weaponName: string;
      equiNumber: number;
    };
  };
  farRange?: {
    [name: string]: {
      name: string;
      set: string;
      slot: string;
      talent: string;
      weaponName: string;
    };
  };
  closeRange?: {
    [name: string]: {
      name: string;
      identifier: string;
      breakFactor: number;
      minBreakFactor: number;
      hand: string;
      shield: string;
      set: string;
      slot: string;
      talent: string;
      weaponName: string;
    };
  };
}